Change sanitization to per field config
The sanitization gem I forked this from would bulk change all fields
based on some system-wide config. So you would say which sanitizations
to apply to a model and it would apply it to all fields (unless marked
otherwise).

My strategy is to more closly match how validations are defined, which
is on a per field basis, each field being able to apply multiple
validations. In like manner, for sanitization you define specific
sanitizers to run on a given field. Each field's sanitizers must be
defined separately.

In the future, ideally it will support custom sanitizers (EachSanitizer)
and sanitize_with which can sanitize the whole model in one go.